+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Setting a preference that doesn't actually change the value won't have any effects (such as MQTT reconnecting) (#1875)
 - Status messages contain the app version and flavour
 - There's now an option to send a remote command "request location" to selected contacts from the map view
+- New experimental preference `discardNetworkLocationThresholdSeconds` that allows OT to discard locations from inaccurate providers (e.g. network) if there's been a recent location from an accurate provider (e.g. gps/fused). (#2053)
 
 ### Bug fixes